feature(SellerService, PostService, SellerController, PostRepository): add search post functionality from user's id
