+++ src/main/java/com/mercadolibre/be_java_hisp_w31_g07/controller/ProductController.java
@@ -1,11 +1,25 @@
 package com.mercadolibre.be_java_hisp_w31_g07.controller;
 
+import com.mercadolibre.be_java_hisp_w31_g07.dto.response.UserPostResponseDto;
 import com.mercadolibre.be_java_hisp_w31_g07.service.IProductService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
+
+import java.util.UUID;
 
 @RestController
 @RequiredArgsConstructor
+@RequestMapping("/products")
 public class ProductController {
     private final IProductService productService;
+
+    @GetMapping("/promo-post/list")
+    public ResponseEntity<UserPostResponseDto> getUserPostDisc(@RequestParam UUID userId){
+        return new ResponseEntity<>(productService.getSellerPromProd(userId), HttpStatus.OK);
+    }
 }
