+++ src/main/java/com/mercadolibre/be_java_hisp_w31_g07/controller/SellerController.java
@@ -51,11 +51,6 @@ public class SellerController {
         return ResponseEntity.ok().build();
     }
 
-    @GetMapping("/users/products/promo-post/list")
-    public ResponseEntity<UserPostResponseDto> getSellerPromProd(@RequestParam UUID userId){
-        return new ResponseEntity<>(sellerService.getSellerPromProd(userId), HttpStatus.OK);
-    }
-
     // FOR TESTING PURPOSES ONLY
     // This endpoint is not part of the original requirements
     // and is only used to verify the functionality of the followSeller method.
