+++ src/main/java/com/mercadolibre/be_java_hisp_w31_g07/service/IProductService.java
@@ -1,7 +1,11 @@
 package com.mercadolibre.be_java_hisp_w31_g07.service;
 
+import com.mercadolibre.be_java_hisp_w31_g07.dto.response.UserPostResponseDto;
+import com.mercadolibre.be_java_hisp_w31_g07.exception.BadRequest;
 import com.mercadolibre.be_java_hisp_w31_g07.model.Product;
 import com.mercadolibre.be_java_hisp_w31_g07.repository.implementations.ProductRepository;
+
+import java.util.UUID;
 
 public interface IProductService {
 
@@ -16,4 +20,13 @@ public interface IProductService {
      * @param product the {@link Product} instance to be created and stored in the repository.
      */
     public void createProduct(Product product);
+
+    /**
+     * Retrieves a seller with a list of post with discount.
+     *
+     * @param userId the unique identifier of the seller to be retrieved
+     * @return a UserPostReponseObject object containing the seller's information and post.
+     * @throws BadRequest if the seller cannot be found
+     */
+    public UserPostResponseDto getSellerPromProd(UUID userId);
 }
