+++ src/main/resources/seller.json
@@ -1,8 +1,6 @@
 [
   {
     "id": "87a3d636-2adf-4640-9872-83e1050427e7",
-    "billingAddress": "123 Calle Ficticia, Ciudad, País",
-    "cuil": "20-12345678-9",
     "followerCount": 0,
     "followers": [
 
@@ -10,8 +8,6 @@
   },
   {
     "id": "928f7404-bd16-43f5-bb3e-25a8c01b68d4",
-    "billingAddress": "12343 Calle Ficticia, Ciudad, País",
-    "cuil": "20-12889678-9",
     "followerCount": 0,
     "followers": [
 
@@ -19,8 +15,6 @@
   },
   {
     "id": "c16c73b4-dcf7-4b8f-9dfb-c421ab3858a4",
-    "billingAddress": "123 Calle Falsa, Ciudad, País",
-    "cuil": "20-12345678-9",
     "followerCount": 0,
     "followers": [
